プラットフォームごとにインポートを分ける
