+++ lib/data/spots.dart
@@ -1,0 +1,14 @@
+import '../models/spot.dart';
+
+const List<Spot> spots = [
+  Spot(
+    title: '五泉市チューリップまつり',
+    location: '五泉市栗本地区一本杉内',
+    image: 'assets/images/tulip_field.jpg',
+    description: '五泉市は、新潟県内でも有数のチューリップの産地として知られています。五泉市では、チューリップの球根を出荷するだけでなく、市民や観光客にも楽しんでもらえるよう、生産者に依頼して巣本地区にチューリップの畑を集め、チューリップまつりを毎年開催しています。最近では、オランダなどから新たな品種の球根を輸入して栽培するなど、多種多様な花が咲き誇っています。',
+    touristTitle: 'ラポルテ五泉',
+    touristLocation: '五泉市赤海863番地',
+    touristDescription: 'ラポルテ五泉は、約2万4,000㎡の敷地に床面積3,700㎡の建物と3つの広場をもつ五泉市の交流拠点複合施設です。  館内には、芸術や学びの場となる多目的ホールや多目的室。市が日本に誇るニットや絹産業、地元の特産物を販売する産直ショップ＆カフェテリア。特におすすめなのは、木造建築を生かして面白遊具を備えた「子どもの遊び場」や開放感あるガレリアの空間です。  約 160 台収容できる駐車場と夜間にも利用できる24時間トイレも備えて車の利用者にも便利です。  "まるっと1日楽しく過ごせる"ラポルテ五泉 みなさまのご利用を、心からお待ちしております。',
+  ),
+  // 他スポットも同様に追加
+];
