+++ lib/utils/api_key_util.dart
@@ -1,0 +1,3 @@
+export 'api_key_util_stub.dart'
+    if (dart.library.html) 'api_key_util_web.dart'
+    if (dart.library.io) 'api_key_util_io.dart';
