+++ lib/utils/api_key_util_io.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_dotenv/flutter_dotenv.dart';
+
+String getGoogleMapsApiKey() {
+  return dotenv.env['GOOGLE_MAPS_API_KEY'] ?? '';
+}
