+++ lib/utils/api_key_util_stub.dart
@@ -1,0 +1,1 @@
+String getGoogleMapsApiKey() => '';
