+++ lib/utils/api_key_util_web.dart
@@ -1,0 +1,5 @@
+import 'dart:html' as html;
+
+String getGoogleMapsApiKey() {
+  return html.window['GOOGLE_MAPS_API_KEY'] as String? ?? '';
+}
